tests: Restore 10-proxy.sh as envoy-smoke-test.sh
[ upstream commit a57e2aad56eedb69079f94c8e1829486623ff9ce ]

A quick sanity check for Envoy integration is still valuable for
day-to-day development with the dev VM. Resgtore the old "10-proxy.sh"
test as "envoy-smoke-test.sh".

Changes to the old script:
- deprecation warnign removed
- saving files on failure commented out
- removal of state on failure commented out
- use of removed '--rev' flag removed
- waiting for identities is more robust

Conflict: the original patch also changes tests/sql-proxy.sh which seems
like an unrelated like a change that crept it, so I optimistically left
it out.

Signed-off-by: Jarno Rajahalme <jarno@covalent.io>
Signed-off-by: Kornilios Kourtis <kornilios@isovalent.com>
Signed-off-by: Robin Hahling <robin.hahling@gw-computing.net>