daemon: Fix handling of policy call map on upgrades
[ upstream commit bb615ea7407ca93e6f094b5da3dd1ad0913ad64b ]

The name of the policy call map on the bpffs was changed between 1.7 and
1.8 by commit 5d6b669. Commit 6bada02 then added code to delete the old
map name on initialization of the agent.

We cannot simply delete the old policy call map because it might still
be used by endpoints (until they are regenerated) and the base programs
(until they are reloaded). However, if we rename the map in the bpffs,
it shouldn't affect BPF programs that are using it and they will then
pick up the new name on reload.

The reverse renaming operation is needed in 1.7 to allow for downgrades
from 1.8.

Fixes: 5d6b669 ("maps/policymap: Rename policy call map to clarify intent")
Fixes: 6bada02 ("daemon: Remove old policy call map")
Signed-off-by: Paul Chaignon <paul@cilium.io>
Signed-off-by: Robin Hahling <robin.hahling@gw-computing.net>