operator: Encase CNP status handover in CLI option
[ upstream commit d1c6a750dabe799fad6bbdc70218d8ab4351c083 ]

The code for handling CNP status updates from other nodes via the
kvstore was previously not covered by the same option that enables this
functionality in the cilium-agent daemon. As such, this could cause the
logic to run, including goroutines for each CNP, in scenarios where this
logic is not in use.

Improve memory usage by disabling this functionality when it is disabled.

Co-authored-by: Joe Stringer <joe@cilium.io>
Signed-off-by: Joe Stringer <joe@cilium.io>
Signed-off-by: André Martins <andre@cilium.io>
Signed-off-by: Kornilios Kourtis <kornilios@isovalent.com>
Signed-off-by: Robin Hahling <robin.hahling@gw-computing.net>