iptables, loader: use interface with default route for EKS rules
[ upstream commit a301853023ef79b36649dc985e0e4b0d9db0edbe ]

Multi-node NodePort traffic on EKS needs specific rules regarding
asymmetric routing. These rules were implemented for the eth0 interface
(namely), because this is what EKS uses. With the default Amazon Linux 2
distribution. But EKS can also run with Ubuntu for example, and the name
of the interface is not the same in that case.

Instead of "eth0", use the interface with the dafault route. This is a
quick fix, and longer term we want to add the rules to all relevant
interfaces, as discussed in #12770.

Fixes: #12770
Fixes: #13143
Signed-off-by: Quentin Monnet <quentin@isovalent.com>
Signed-off-by: Michal Rostecki <mrostecki@opensuse.org>
Signed-off-by: Robin Hahling <robin.hahling@gw-computing.net>