ipam: Warn when failing to update CN status
[ upstream commit a7451f82d4c6ed96a9e94d608ec71f4255d55c65 ]

This commit attempts to improve the readability and debugability of
(*Node).syncToAPIServer(). The main motivation for this change is to fix
the swallowing of the error when UpdateStatus fails, and the subsequent
Get succeeds.

To that end, this commit adds a warning log msg and propagates the error
in case we never succeed, to surface any trouble with updating the
status of the CiliumNode (CN) resource. Failures in updating the status
can be indicative of general connectivity problems in ENI and Azure IPAM
modes. As seen in https://github.com/cilium/cilium/issues/13193, it can
also lead to Cilium never becoming ready because the agent was never
able to retrieve ENI information required to make allocations, etc. (The
former was caused by the Operator being blocked from "publishing" the
ENI information into the CN resource, which is why the agent was unable
to retrieve it.)

Signed-off-by: Chris Tarazi <chris@isovalent.com>
Signed-off-by: Kornilios Kourtis <kornilios@isovalent.com>
Signed-off-by: Robin Hahling <robin.hahling@gw-computing.net>