daemon: listen on IPv4 and IPv6 for health endpoint
[ upstream commit 02a36119627b5d7cd5fa560305716a5ba64b640c ]

If the agent liveness/readiness probe host is set to the IPv6 address
::1 instead of the default IPv4 127.0.0.1, Cilium never becomes ready in
an IPv6-only environment. This is because the daemon health endpoint
currently listens on localhost:9876 which will not listen on both IPv4
and IPv6.

To fix this, listen on both IPv4 and IPv6 explicitly (depending on the
daemons's tenable-ipv{4,6} flags) and only fail with an error if both of
them fail or one was disabled and the other one fails.

Fixes #13165

Signed-off-by: Tobias Klauser <tklauser@distanz.ch>
Signed-off-by: Michal Rostecki <mrostecki@opensuse.org>
Signed-off-by: Robin Hahling <robin.hahling@gw-computing.net>