hubble: Fix filter by reply reporting flows with unknown reply state
[ upstream commit 511b15d0e53cbe311465c624adbf9bc3fe492592 ]

This fixes a bug in the reply filter on `reply=false` would report flows
for which we actually do not know if they were replies or not. Not all
trace points have connection tracking state available, thus looking at
the reply flag alone is not sufficent to tell if something a flow was a
reply or not.

Ideally, we would fix this in the parser and make the `reply` an
optional boolean, so we can distinguish between a `false` value and an
absent value. This however is a breaking change in the Hubble API, which
we want to avoid. Therefore, this commit modifies the reply filter to
only report flows here for which we know that the reply field is
reliable.

Signed-off-by: Sebastian Wicki <sebastian@isovalent.com>
Signed-off-by: Michal Rostecki <mrostecki@opensuse.org>
Signed-off-by: Robin Hahling <robin.hahling@gw-computing.net>