k8s: Consider session affinity parameters when comparing Services
[ upstream commit 745a26b2ead09e689f8626413467d4083a7fa00a ]

Previously, if the `sessionAffinity` or `sessionAffinityTimeoutSec` of a Service
changed, cilium would ignore the change. This meant that even though cilium
correctly handles `sessionAffinity`, changing the `sessionAffinity` of a service
without any other changes did not cause cilium to change its handling.

Add `sessionAffinity` and `sessionAffinityTimeoutSec` handling to the comparator
for services, so that cilium will notice affinity changes.

Signed-off-by: Adam Wolfe Gordon <awg@digitalocean.com>
Signed-off-by: Robin Hahling <robin.hahling@gw-computing.net>