clusterpool: Error out if IPv6 is enabled but ClusterPool CIDR is not set
[ upstream commit 5ce94ef676fada7d4d59d180e62c815935d13f11 ]

The existing logic reported an error if both IPv4 and IPv6 CIDRs were
not provided. This is insufficient when both IPv4 and IPv6 are enabled
but only an IPv4 ClusterPool CIDR is specified.

It lead to a situation where agents won't start up while waiting for the
CiliumNode to to provide an IPv6 address.

Signed-off-by: Thomas Graf <thomas@cilium.io>
Signed-off-by: Quentin Monnet <quentin@isovalent.com>
Signed-off-by: Sebastian Wicki <sebastian@isovalent.com>